# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Gaming Companion
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameDetectionConfidence = typeof GameDetectionConfidence[keyof typeof GameDetectionConfidence];


export const GameDetectionConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
  none: 'none',
} as const;

export interface GameDetection {
  detected: boolean;
  /** @nullable */
  gameName: string | null;
  /** @nullable */
  processName: string | null;
  confidence: GameDetectionConfidence;
}

export interface Screenshot {
  /**
     * Base64 encoded PNG image data
     * @nullable
     */
  imageData: string | null;
  /**
     * ISO timestamp when screenshot was taken
     * @nullable
     */
  capturedAt: string | null;
  available: boolean;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  message: string;
  /**
     * Currently detected game name for context
     * @nullable
     */
  gameName?: string | null;
  /** Whether to attach the latest screenshot to the message */
  includeScreenshot?: boolean;
}

export interface ChatResponse {
  reply: string;
  model: string;
  tokensUsed: number;
}

export interface Settings {
  /** Whether a Claude API key is configured (never returns the actual key) */
  hasApiKey: boolean;
  /** Auto-screenshot interval in seconds */
  screenshotInterval: number;
  /** Whether auto screenshot capture is enabled */
  autoCapture: boolean;
}

export interface SettingsInput {
  /**
     * Claude API key (pass null to keep existing)
     * @nullable
     */
  apiKey?: string | null;
  /**
     * Auto-screenshot interval in seconds (10-300)
     * @nullable
     */
  screenshotInterval?: number | null;
  /** @nullable */
  autoCapture?: boolean | null;
}

export interface ErrorResponse {
  error: string;
}