# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for AI Gaming Companion
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message to the AI assistant
      description: Sends a user message along with game context and optional screenshot to Claude
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request (e.g. missing API key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        gameName:
          type: ["string", "null"]
          description: Currently detected game name for context
        includeScreenshot:
          type: boolean
          description: Whether to attach the latest screenshot to the message
      required:
        - message
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        model:
          type: string
        tokensUsed:
          type: integer
      required:
        - reply
        - model
        - tokensUsed
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error