# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Gaming Companion
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  /** @minLength 1 */
  message: string;
  /**
     * Currently detected game name for context
     * @nullable
     */
  gameName?: string | null;
  /** Whether to attach the latest screenshot to the message */
  includeScreenshot?: boolean;
}